feat(target): add transformation creating topological order of entities
For Python, and maybe other languages, we need to create a system of
ordering the entities, such that for any two entities (A, B), if A
depends on B, then A will come after B in the entities-list.

The DependencyWalker transformation visits every entity in the model and
creates a graph where an edge in the graph (A, B) symbolises that A
depends on B. The `topological_sort()` method creates a topological
order of the graph and overwrites the model entities with the newly
sorted list of entities. The method throws an error if there are
circular dependencies.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>